requirements.txt
altgraph==0.16.1
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
future==0.17.1
idna==2.8
macholib==1.11
pefile==2018.8.8
PyInstaller==3.4
pythonnet==2.3.0
pywebview==2.2.1
pywin32-ctypes==0.2.0
requests==2.21.0
urllib3==1.24.1
